+++ CHANGELOG.md
@@ -28,6 +28,7 @@ _ `_optional` subpackage for managing optional dependencies
   `DiscreteCardinalityConstraint`/`ContinuousCardinalityConstraint` subclasses
 - Uniform sampling mechanism for continuous spaces with cardinality constraints
 - Enable `ContinuousCardinalityConstraint` in `BotorchRecommender`
+- Attribute `n_threshold_inactive_parameters_generator` added to `BotorchRecommender`  
 - Properties `combinatorial_zero_parameters` and 
   `combinatorial_counts_zero_parameters` in both `ContinuousCardinalityConstraint`and 
   `SubspaceContinuous`